TODO: Don't care about weird table header background colours.
It seems you can reproduce this by searching for stuff, selecting all and scrolling down a
bit (notice how the headers are a bit transparent) and then searching for something else which
clears the selection, but doesn't seem to fix the header colour.  Oh well, i am choosing not to care
about this.