Restructure loading of ~/.config/spotiqueue/init.scm.
The old way meant that all of the user's Scheme was executed _before_ the UI was ready.  This isn't
nice if we want to be able to (for example) `(player:set-auto-advance #f)` in our startup script.
So, let's only load the init.scm once we've received applicationDidFinishLoading.  It's a bit more
opaque for the user, maybe, but the bundled init.scm file wasn't really intended to be edited anyway.