init.scm(define-key): Procedures instead of symbols in keymaps.
Storing symbols such as `'queue:delete-selected-tracks` in the keybinding hashmap works great if
everything runs in one big module (as it used to, in `spotiqueue init`), but if we want to let the
user run their own config as a module things get tricky, since symbols won't necessarily be in-scope
in that module.  We could hack around that by forcing the user to import `spotiqueue internal` even
if they don't intend to do anything with it, but that seems icky.

A nicer solution seems to be to depart from the Emacsy approach and simply store procedure
references directly in the hashmap.  This way when we evaluate the contents of the hashmap, we don't
have to cross our fingers the symbol is in-scope -- Guile will chase it up for us.

* guile/spotiqueue/init.scm (define-key): Expect procedure as `action` argument.
* examples/paul-config.scm: Remove `spotiqueue internal` from use-modules list.