RBGuileBridge.swift(guile_handle_key): Remove `scm_eval`.
We don't need the `eval` call any more since we don't expect symbols but procedures in the keymap.
This also avoids the situation where doing `eval` in the `(current-module)` context leads us
astray.

I guess this boils down to a departure from the Emacs keybinding style where you can bind a key to a
symbol 'action so that you get late-binding to whatever `action` happens to be when you hit the key,
to our approach where the expectation is that the binding is early but can happily be overwritten by
a subsequent call to `define-key`. 🤷‍♀️